Comply with BSD-3 license of `temporary` package
Author's note: "stolen from" is not a sufficient legal disclaimer.

The BSD-3 license requires a copy of the license in source & in
documentation for any derivative work, which `UnliftIO.Temporary` is by
its own admission. This places that copy in the module Haddocks so it's
supplied in both the source and (collapsed) documentation.